requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2
fake-useragent==0.1.11
idna==2.10
img2pdf==0.4.0
lxml==4.6.3
numpy==1.20.2
pandas==1.2.4
pikepdf==2.11.1
Pillow==8.2.0
PyPDF2==1.26.0
pyTelegramBotAPI==3.7.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
schedule==1.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4